Crate wlroots

This crate provides safe bindings to wlroots.

Re-exports

pub extern crate wlroots_sys;

pub extern crate xkbcommon;

pub use wlroots_dehandle::wlroots_dehandle;

pub use wlroots_sys::wl_shm_format::*;

pub use wlroots_sys::wlr_axis_orientation::*;

pub use wlroots_sys::wlr_axis_source::*;

pub use wlroots_sys::wlr_button_state::*;

pub use wlroots_sys::wlr_input_device_type::*;

pub use wlroots_sys::wlr_key_state::*;

pub use wlroots_sys::wlr_keyboard_modifier::*;

pub use wlroots_sys::wlr_tablet_pad_ring_source::*;

pub use wlroots_sys::wlr_tablet_pad_strip_source::*;

pub use wlroots_sys::wlr_tablet_tool_proximity_state::*;

Modules

area
Wrapper for the wlr_box type. Note that we renamed it to Area to avoid conflicts with Rust's Box.

backend

compositor
Main entry point to the library. See examples for documentation on how to use this struct.

cursor

data_device

extensions

input

output

render

seat

shell

surface

utils
Utilities for use within wlroots-rs that are not directly related to Wayland or compositors.

xwayland

Macros

with_handles
A convenience macro designed for use with Handle types.

wlr_log
Logs a message using wlroots' logging capability.

Enums

TabletToolAxes

wl_shm_format

wlr_axis_orientation

wlr_axis_source

wlr_button_state

wlr_input_device_type

wlr_key_state

wlr_keyboard_modifier

wlr_tablet_pad_ring_source

wlr_tablet_pad_strip_source

wlr_tablet_tool_proximity_state