Module wlroots::backend

Structs

Device

Drm
When the compositor is ran on a TTY and has full control of the system resources.

Headless
In this backend the only resource the compositor uses is the Wayland file descriptor. It doesn't try to grab actual keyboard/pointers and it doesn't render anything.

Libinput

Multi
When multiple backends are running or when the compositor writer doesn't care and just used the auto create option in the CompositorBuilder.

Session

Wayland
When the compositor is running in a nested Wayland environment. e.g. your compositor is executed while the user is running Gnome+Mutter or Weston.

X11
When the compositor is running in a nested X11 environment. e.g. your compositor is executed while the user is running an X11 window manager.

Enums

Backend

Type Definitions

UnsafeRenderSetupFunction
A custom function to set up the renderer.