Module wlroots::input::pointer::event

Pointers and their events

Structs

AbsoluteMotion
Event that triggers when data from a device that supports absolute motion sends data to the compositor.

Axis
Event that triggers when the pointer device scrolls (e.g using a wheel

Button
Event that triggers when a button is pressed (e.g left click, right click, a gaming mouse button, etc.).

Motion
Event that triggers when the pointer moves.

Constants

BTN_BACK

BTN_EXTRA

BTN_FORWARD

BTN_LEFT

BTN_MIDDLE

BTN_MOUSE

BTN_RIGHT

BTN_SIDE

BTN_TASK