Trait wlroots::input::touch::Handler

pub trait Handler {
    fn on_down(
        &mut self, 
        compositor_handle: Handle, 
        touch_handle: Handle, 
        event: &Down
    ) { ... }

    fn on_up(
        &mut self, 
        compositor_handle: Handle, 
        touch_handle: Handle, 
        event: &Up
    ) { ... }

    fn on_motion(
        &mut self, 
        compositor_handle: Handle, 
        touch_handle: Handle, 
        event: &Motion
    ) { ... }

    fn on_cancel(
        &mut self, 
        compositor_handle: Handle, 
        touch_handle: Handle, 
        event: &Cancel
    ) { ... }

    fn destroyed(&mut self, compositor_handle: Handle, touch_handle: Handle) { ... }
}

Provided Methods

fn on_down(
&mut self,
compositor_handle: Handle,
touch_handle: Handle,
event: &Down
)

Callback that is triggered when the user starts touching the screen/input device.

fn on_up(&mut self, compositor_handle: Handle, touch_handle: Handle, event: &Up)

Callback that is triggered when the user stops touching the screen/input device.

fn on_motion(
&mut self,
compositor_handle: Handle,
touch_handle: Handle,
event: &Motion
)

Callback that is triggered when the user moves his fingers along the screen/input device.

fn on_cancel(
&mut self,
compositor_handle: Handle,
touch_handle: Handle,
event: &Cancel
)

Callback triggered when the touch is canceled.

fn destroyed(&mut self, compositor_handle: Handle, touch_handle: Handle)

Callback that is triggered when the touch is destroyed.