Module wlroots::output

Modules

layout
TODO Documentation

manager
Output resources are managed by the output resource manager.

Structs

Builder
Used to ensure the output sets the mode before doing any other operation on the Output.

BuilderResult
Used to ensure that the builder is used to construct the output::Handler instance.

Cursor

Damage
Tracks damage for an output.

Mode

Output

Traits

Handler

Type Definitions

Handle

Subpixel

Transform