Struct wlroots::output::Output

pub struct Output { /* fields omitted */ }

Methods

impl Output

pub fn choose_best_mode(&mut self)

Sets the best modesetting for an output.

NOTE You cannot call this when the output will be removed.

I'm still marking it as safe though because we protect against that action in the output destruction callback.

pub fn set_mode(&mut self, mode: Mode) -> bool

Set this to be the current mode for the Output.

pub fn set_custom_mode(&mut self, size: Size, refresh: i32) -> bool

Set a custom mode for this output.

pub fn name(&self) -> String

Gets the name of the output in UTF-8.

pub fn make(&self) -> String

Gets the make of the output in UTF-8.

pub fn model(&self) -> String

Gets the model of the output in UTF-8.

pub fn serial(&self) -> String

Gets the serial of the output in UTF-8.

pub fn enabled(&self) -> bool

Determines if the output is enabled or not.

pub fn scale(&self) -> c_float

Get the scale of the output

pub fn needs_swap(&self) -> bool

Determines if the output should have its buffers swapped or not.

pub fn refresh_rate(&self) -> i32

Get the refresh rate of the output.

pub fn current_mode<'output>(&'output self) -> Option<Mode<'output>>

pub fn layout_space_pos(&self) -> (i32, i32)

Gets the output position in layout space reported to clients.

pub fn subpixel(&self) -> Subpixel

Get subpixel information about the output.

pub fn get_transform(&self) -> Transform

Get the transform information about the output.

pub fn schedule_frame(&mut self)

Manually schedules a frame event.

If a frame event is already pending, it is a no-op.

pub unsafe fn make_current(&mut self) -> (bool, Option<c_int>)

Make this output the current output.

Unsafety

This is done for rendering purposes, and you should really use a GenericRenderer instead in order to do this.

Sometimes however you need to do e.g opengl rendering and we haven't wrapped that. If that's the case, call this first and then swap the buffers.

Returns the drawing buffer age in number of frames in number of frames, or None if unknown. This is useful for damage tracking.

pub unsafe fn swap_buffers<'a, T, U>(&mut self, when: T, damage: U) -> bool where
T: Into<Option<Duration>>,
U: Into<Option<&'a mut PixmanRegion>>,

Swaps the buffers and draws whatever is in the back buffer on the screen.

If the time of the frame is not known, set when to None.

If the compositor does not support damage tracking, set damage to None

Unsafety

This is done for rendering purposes, but if called multiple times then you could cause a deadlock.

You should try to use a GenericRenderer, but sometimes it's necessary to do your own manual rendering in a compositor. In that case, call make_current, do your rendering, and then call this function.

pub fn frame_pending(&self) -> bool

Determines if a frame is pending or not.

pub fn size(&self) -> (i32, i32)

Get the dimensions of the output as (width, height).

pub fn physical_size(&self) -> (i32, i32)

Get the physical dimensions of the output as (width, height).

pub fn transformed_resolution(&self) -> (c_int, c_int)

Computes the transformed output resolution

pub fn effective_resolution(&self) -> (c_int, c_int)

Computes the transformed and scaled output resolution.

pub fn transform_matrix(&self) -> [c_float; 9]

pub fn transform(&mut self, transform: Transform)

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn modes<'output>(&'output self) -> Vec<Mode<'output>>

Get the modes associated with this output.

Note that some backends may have zero modes.

pub fn enable(&mut self, enable: bool) -> bool

Enables or disables an output.

pub fn set_gamma(&mut self, size: usize, r: u16, g: u16, b: u16) -> bool

Sets the gamma based on the size.

pub fn get_gamma_size(&self) -> usize

Get the gamma size.

pub fn set_position(&mut self, origin: Origin)

Sets the position of this output.

pub fn set_scale(&mut self, scale: c_float)

Set the scale applied to this output.

pub fn damage(&mut self) -> &mut Damage