Module wlroots::render

Modules

matrix
Matrix math is used to render things on a computer screen. This is common throughout computer graphics, for examples and primers on using matrix math to render things to screens please read an OpengGL tutorial.

Structs

GenericRenderer
A generic interface for rendering to the screen.

Image
An image that can be attached to a Cursor or OutputCursor.

PixmanRegion
A pixman region, used for damage tracking.

Renderer
The state machine type that allows you to manipulate a screen and its buffer.

Texture
A wrapper for a wlr_texture.

Enums

TextureFormat
Wrapper around wl_shm_format, to make it easier and nicer to type.