Trait wlroots::seat::Handler

pub trait Handler {
    fn pointer_grabbed(
        &mut self, 
        compositor_handle: Handle, 
        seat_handle: Handle, 
        event: &Pointer
    ) { ... }

    fn pointer_released(
        &mut self, 
        compositor_handle: Handle, 
        seat_handle: Handle, 
        event: &Pointer
    ) { ... }

    fn keyboard_grabbed(
        &mut self, 
        compositor_handle: Handle, 
        seat_handle: Handle, 
        event: &Keyboard
    ) { ... }

    fn keyboard_released(
        &mut self, 
        compositor_handle: Handle, 
        seat_handle: Handle, 
        event: &Keyboard
    ) { ... }

    fn touch_grabbed(
        &mut self, 
        compositor_handle: Handle, 
        seat_handle: Handle, 
        event: &Touch
    ) { ... }

    fn touch_released(
        &mut self, 
        compositor_handle: Handle, 
        seat_handle: Handle, 
        event: &Touch
    ) { ... }

    fn cursor_set(
        &mut self, 
        compositor_handle: Handle, 
        seat_handle: Handle, 
        event: &SetCursor
    ) { ... }

    fn received_selection(
        &mut self, 
        compositor_handle: Handle, 
        seat_handle: Handle
    ) { ... }

    fn primary_selection(
        &mut self, 
        compositor_handle: Handle, 
        seat_handle: Handle
    ) { ... }

    fn destroy(&mut self, compositor_handle: Handle, seat_handle: Handle) { ... }

    fn new_drag_icon(
        &mut self, 
        compositor_handle: Handle, 
        seat_handle: Handle, 
        Handle
    ) -> (Option<Box<dyn Handler>>, Option<Box<dyn Handler>>) { ... }
}

Provided Methods

fn pointer_grabbed(
&mut self,
compositor_handle: Handle,
seat_handle: Handle,
event: &Pointer
)

Callback triggered when a client has grabbed a pointer.

fn pointer_released(
&mut self,
compositor_handle: Handle,
seat_handle: Handle,
event: &Pointer
)

Callback triggered when a client has ended a pointer grab.

fn keyboard_grabbed(
&mut self,
compositor_handle: Handle,
seat_handle: Handle,
event: &Keyboard
)

Callback triggered when a client has grabbed a keyboard.

fn keyboard_released(
&mut self,
compositor_handle: Handle,
seat_handle: Handle,
event: &Keyboard
)

Callback triggered when a client has ended a keyboard grab.

fn touch_grabbed(
&mut self,
compositor_handle: Handle,
seat_handle: Handle,
event: &Touch
)

Callback triggered when a client has grabbed a touch.

fn touch_released(
&mut self,
compositor_handle: Handle,
seat_handle: Handle,
event: &Touch
)

Callback triggered when a client has ended a touch grab.

fn cursor_set(
&mut self,
compositor_handle: Handle,
seat_handle: Handle,
event: &SetCursor
)

Callback triggered when a client sets the cursor for this seat.

E.g this happens when the seat enters a surface.

fn received_selection(&mut self, compositor_handle: Handle, seat_handle: Handle)

The seat was provided with a selection by the client.

fn primary_selection(&mut self, compositor_handle: Handle, seat_handle: Handle)

The seat was provided with a selection from the primary buffer by the client.

fn destroy(&mut self, compositor_handle: Handle, seat_handle: Handle)

The seat is being destroyed.

fn new_drag_icon(
&mut self,
compositor_handle: Handle,
seat_handle: Handle,
Handle
) -> (Option<Box<dyn Handler>>, Option<Box<dyn Handler>>)

A new drag icon has been created.