Struct wlroots::shell::xdg_shell::Surface

pub struct Surface { /* fields omitted */ }

Methods

impl Surface

pub fn surface(&mut self) -> Handle

Gets the surface used by this XDG shell.

pub fn role(&self) -> wlr_xdg_surface_role

Get the role of this XDG surface.

pub fn state(&mut self) -> Option<&mut ShellState>

pub fn configured(&self) -> bool

Determines if this XDG shell surface has been configured or not.

pub fn added(&self) -> bool

pub fn configure_serial(&self) -> u32

pub fn configure_next_serial(&self) -> u32

pub fn has_next_geometry(&self) -> bool

pub fn next_geometry(&self) -> Area

pub fn geometry(&self) -> Area

pub fn ping(&mut self)

Send a ping to the surface.

If the surface does not respond with a pong within a reasonable amount of time, the ping timeout event will be emitted.

pub fn surface_at(
&mut self,
sx: f64,
sy: f64,
sub_sx: &mut f64,
sub_sy: &mut f64
) -> Option<Handle>

Find a surface within this surface at the surface-local coordinates.

Returns the popup and coordinates in the topmost surface coordinate system or None if no popup is found at that location.

pub fn for_each_surface<F>(&self, iterator: F) where
F: FnMut(Handle, i32, i32),