Trait wlroots::shell::xdg_shell::Handler

pub trait Handler {
    fn on_commit(
        &mut self, 
        compositor_handle: Handle, 
        surface_handle: Handle, 
        xdg_shell_handle: Handle
    ) { ... }

    fn destroyed(&mut self, Handle, Handle) { ... }

    fn ping_timeout(
        &mut self, 
        compositor_handle: Handle, 
        surface_handle: Handle, 
        xdg_shell_handle: Handle
    ) { ... }

    fn new_popup(
        &mut self, 
        compositor_handle: Handle, 
        surface_handle: Handle, 
        xdg_shell_handle: Handle
    ) { ... }

    fn maximize_request(
        &mut self, 
        compositor_handle: Handle, 
        surface_handle: Handle, 
        xdg_shell_handle: Handle
    ) { ... }

    fn minimize_request(
        &mut self, 
        compositor_handle: Handle, 
        surface_handle: Handle, 
        xdg_shell_handle: Handle
    ) { ... }

    fn move_request(
        &mut self, 
        compositor_handle: Handle, 
        surface_handle: Handle, 
        xdg_shell_handle: Handle, 
        event: &Move
    ) { ... }

    fn resize_request(
        &mut self, 
        compositor_handle: Handle, 
        surface_handle: Handle, 
        xdg_shell_handle: Handle, 
        event: &Resize
    ) { ... }

    fn fullscreen_request(
        &mut self, 
        compositor_handle: Handle, 
        surface_handle: Handle, 
        xdg_shell_handle: Handle, 
        event: &SetFullscreen
    ) { ... }

    fn show_window_menu_request(
        &mut self, 
        compositor_handle: Handle, 
        surface_handle: Handle, 
        xdg_shell_handle: Handle, 
        event: &ShowWindowMenu
    ) { ... }

    fn map_request(
        &mut self, 
        compositor_handle: Handle, 
        surface_handle: Handle, 
        xdg_shell_handle: Handle
    ) { ... }

    fn unmap_request(
        &mut self, 
        compositor_handle: Handle, 
        surface_handle: Handle, 
        xdg_shell_handle: Handle
    ) { ... }
}

Handles events from the client stable XDG shells.

Provided Methods

fn on_commit(
&mut self,
compositor_handle: Handle,
surface_handle: Handle,
xdg_shell_handle: Handle
)

Called when the surface recieve a request event.

fn destroyed(&mut self, Handle, Handle)

Called when the wayland shell is destroyed (e.g by the user)

fn ping_timeout(
&mut self,
compositor_handle: Handle,
surface_handle: Handle,
xdg_shell_handle: Handle
)

Called when the ping request timed out.

This usually indicates something is wrong with the client.

fn new_popup(
&mut self,
compositor_handle: Handle,
surface_handle: Handle,
xdg_shell_handle: Handle
)

Called when a new popup appears in the xdg tree.

fn maximize_request(
&mut self,
compositor_handle: Handle,
surface_handle: Handle,
xdg_shell_handle: Handle
)

Called when there is a request to maximize the XDG surface.

fn minimize_request(
&mut self,
compositor_handle: Handle,
surface_handle: Handle,
xdg_shell_handle: Handle
)

Called when there is a request to minimize the XDG surface.

fn move_request(
&mut self,
compositor_handle: Handle,
surface_handle: Handle,
xdg_shell_handle: Handle,
event: &Move
)

Called when there is a request to move the shell surface somewhere else.

fn resize_request(
&mut self,
compositor_handle: Handle,
surface_handle: Handle,
xdg_shell_handle: Handle,
event: &Resize
)

Called when there is a request to resize the shell surface.

fn fullscreen_request(
&mut self,
compositor_handle: Handle,
surface_handle: Handle,
xdg_shell_handle: Handle,
event: &SetFullscreen
)

Called when there is a request to make the shell surface fullscreen.

fn show_window_menu_request(
&mut self,
compositor_handle: Handle,
surface_handle: Handle,
xdg_shell_handle: Handle,
event: &ShowWindowMenu
)

Called when there is a request to show the window menu.

fn map_request(
&mut self,
compositor_handle: Handle,
surface_handle: Handle,
xdg_shell_handle: Handle
)

Called when the surface is ready to be mapped. It should be added to the list of views at this time.

fn unmap_request(
&mut self,
compositor_handle: Handle,
surface_handle: Handle,
xdg_shell_handle: Handle
)

Called when the surface should be unmapped. It should be removed from the list of views at this time, but may be remapped at a later time.