Struct wlroots::xwayland::surface::Surface

pub struct Surface { /* fields omitted */ }

An Xwayland user interface component. It has an absolute position in layout-local coordinates.

When a surface is ready to be displayed, the map event is emitted. When a surface should no longer be displayed, the unmap event is emitted.

The unmap event is guaranteed to be emitted before the destroy event if the view is destroyed when mapped.

Methods

impl Surface

pub fn window_id(&self) -> xcb_window_t

Get the window id for this surface.

pub fn surface_id(&self) -> u32

Get the surface id for this surface.

pub fn surface(&self) -> Option<Handle>

Get the Wayland surface associated with this Surface. If the shell surface is not mapped, then it has no surface, and this will return None.

pub fn coords(&self) -> (int16_t, int16_t)

Get the coordinates of the window.

Return format is (x, y)

pub fn dimensions(&self) -> (uint16_t, uint16_t)

Get the dimensions the XWayland surface.

Return format is (width, height).

pub fn saved_dimensions(&self) -> (uint16_t, uint16_t)

TODO What does this represent?

Return format is (width, height)

pub fn override_redirect(&self) -> bool

TODO What does this represent?

pub fn mapped(&self) -> bool

pub fn title(&self) -> Option<String>

Get the title of the client, if there is one.

pub fn class(&self) -> Option<String>

Get the class of the client, if there is one.

pub fn instance(&self) -> Option<String>

Get the instance of the client, if there is one.

pub fn pid(&self) -> pid_t

Get the PID associated with the client.

pub fn parent(&self) -> Option<Handle>

Get the parent surface if there is one.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn children(&self) -> Vec<Handle>

Get the list of children surfaces.

pub unsafe fn window_type(&self) -> *mut xcb_atom_t

Get the type of the window from xcb.

pub unsafe fn window_type_len(&self) -> size_t

Get the length of the window_type ptr

pub unsafe fn protocols(&self) -> *mut xcb_atom_t

Get the protocols of the client.

pub unsafe fn protocols_len(&self) -> size_t

Get the length of the protocols ptr.

pub fn decorations(&self) -> u32

Get the decorations on this XWayland client.

pub fn hints<'surface>(&'surface self) -> Hints<'surface>

Get any surface hints the client is providing.

pub fn size_hints<'surface>(&'surface self) -> SizeHints<'surface>

Get any size hints the client is providing.

pub fn hints_urgency(&self) -> u32

Get the urgency of the hints.

pub fn pinging(&self) -> bool

pub unsafe fn ping_timer(&self) -> *mut wl_event_source

pub fn fullscreen(&self) -> bool

Determine if the client is fullscreen or not.

pub fn maximized_vert(&self) -> bool

Determine if the client is maximized vertically.

pub fn maximized_horz(&self) -> bool

Determine if the client is maximized horizontally.

pub fn has_alpha(&self) -> bool

Determine if the client has an alpha channel.

pub fn geometry(&self) -> Area

Geometry of the surface in layout-local coordinates

pub fn configure(&self, x: i16, y: i16, width: u16, height: u16)

Send the surface a configure request, requesting the new position and dimensions

pub fn set_activated(&self, active: bool)

Tell the window whether it is the foucsed window