Trait wlroots::xwayland::surface::Handler

pub trait Handler {
    fn destroyed(
        &mut self, 
        compositor_handle: Handle, 
        surface_handle: Option<Handle>, 
        xwayland_surface_handle: Handle
    ) { ... }

    fn on_configure(
        &mut self, 
        compositor_handle: Handle, 
        surface_handle: Option<Handle>, 
        xwayland_surface_handle: Handle, 
        configure: &Configure
    ) { ... }

    fn on_move(
        &mut self, 
        compositor_handle: Handle, 
        surface_handle: Option<Handle>, 
        xwayland_surface_handle: Handle, 
        event: &Move
    ) { ... }

    fn on_resize(
        &mut self, 
        compositor_handle: Handle, 
        surface_handle: Option<Handle>, 
        xwayland_surface_handle: Handle, 
        event: &Resize
    ) { ... }

    fn on_maximize(
        &mut self, 
        compositor_handle: Handle, 
        surface_handle: Option<Handle>, 
        xwayland_surface_handle: Handle
    ) { ... }

    fn on_fullscreen(
        &mut self, 
        compositor_handle: Handle, 
        surface_handle: Option<Handle>, 
        xwayland_surface_handle: Handle
    ) { ... }

    fn on_map(
        &mut self, 
        compositor_handle: Handle, 
        surface_handle: Option<Handle>, 
        xwayland_surface_handle: Handle
    ) -> Option<Box<dyn Handler>> { ... }

    fn on_unmap(
        &mut self, 
        compositor_handle: Handle, 
        surface_handle: Option<Handle>, 
        xwayland_surface_handle: Handle
    ) { ... }

    fn title_set(
        &mut self, 
        compositor_handle: Handle, 
        surface_handle: Option<Handle>, 
        xwayland_surface_handle: Handle
    ) { ... }

    fn class_set(
        &mut self, 
        compositor_handle: Handle, 
        surface_handle: Option<Handle>, 
        xwayland_surface_handle: Handle
    ) { ... }

    fn parent_set(
        &mut self, 
        compositor_handle: Handle, 
        surface_handle: Option<Handle>, 
        xwayland_surface_handle: Handle
    ) { ... }

    fn pid_set(
        &mut self, 
        compositor_handle: Handle, 
        surface_handle: Option<Handle>, 
        xwayland_surface_handle: Handle
    ) { ... }

    fn window_type_set(
        &mut self, 
        compositor_handle: Handle, 
        surface_handle: Option<Handle>, 
        xwayland_surface_handle: Handle
    ) { ... }

    fn ping_timeout(
        &mut self, 
        compositor_handle: Handle, 
        surface_handle: Option<Handle>, 
        xwayland_surface_handle: Handle
    ) { ... }
}

Provided Methods

fn destroyed(
&mut self,
compositor_handle: Handle,
surface_handle: Option<Handle>,
xwayland_surface_handle: Handle
)

Called when the XWayland surface is destroyed (e.g by the user).

fn on_configure(
&mut self,
compositor_handle: Handle,
surface_handle: Option<Handle>,
xwayland_surface_handle: Handle,
configure: &Configure
)

Called when the XWayland surface wants to be configured.

fn on_move(
&mut self,
compositor_handle: Handle,
surface_handle: Option<Handle>,
xwayland_surface_handle: Handle,
event: &Move
)

Called when the XWayland surface wants to move.

fn on_resize(
&mut self,
compositor_handle: Handle,
surface_handle: Option<Handle>,
xwayland_surface_handle: Handle,
event: &Resize
)

Called when the XWayland surface wants to be resized.

fn on_maximize(
&mut self,
compositor_handle: Handle,
surface_handle: Option<Handle>,
xwayland_surface_handle: Handle
)

Called when the XWayland surface wants to be maximized.

fn on_fullscreen(
&mut self,
compositor_handle: Handle,
surface_handle: Option<Handle>,
xwayland_surface_handle: Handle
)

Called when the XWayland surface wants to be fullscreen.

fn on_map(
&mut self,
compositor_handle: Handle,
surface_handle: Option<Handle>,
xwayland_surface_handle: Handle
) -> Option<Box<dyn Handler>>

fn on_unmap(
&mut self,
compositor_handle: Handle,
surface_handle: Option<Handle>,
xwayland_surface_handle: Handle
)

fn title_set(
&mut self,
compositor_handle: Handle,
surface_handle: Option<Handle>,
xwayland_surface_handle: Handle
)

Called when the title has been set on the XWayland surface.

fn class_set(
&mut self,
compositor_handle: Handle,
surface_handle: Option<Handle>,
xwayland_surface_handle: Handle
)

Called when the class has been set on the XWayland surface.

fn parent_set(
&mut self,
compositor_handle: Handle,
surface_handle: Option<Handle>,
xwayland_surface_handle: Handle
)

Called when the parent has been set on the XWayland surface.

fn pid_set(
&mut self,
compositor_handle: Handle,
surface_handle: Option<Handle>,
xwayland_surface_handle: Handle
)

Called when the PID has been set on the XWayland surface.

fn window_type_set(
&mut self,
compositor_handle: Handle,
surface_handle: Option<Handle>,
xwayland_surface_handle: Handle
)

Called when the window type has been set on the XWayland surface.

fn ping_timeout(
&mut self,
compositor_handle: Handle,
surface_handle: Option<Handle>,
xwayland_surface_handle: Handle
)

Called when the ping request timed out.

This usually indicates something is wrong with the client.